+++ torch/csrc/autograd/functions/init.cpp
@@ -1,4 +1,3 @@
-#include "torch/csrc/utils/pybind.h"
 #include "batch_normalization.h"
 #include "convolution.h"
 #include "accumulate_grad.h"
@@ -6,31 +5,10 @@
 #include "tensor.h"
 #include "special.h"
 #include "jit_closure.h"
-#include "torch/csrc/THP.h"
-#include "torch/csrc/autograd/python_cpp_function.h"
+#include "torch/csrc/autograd/functions/pybind.h"
 #include "torch/csrc/jit/python_tracer.h"
+#include "torch/csrc/utils/pybind.h"
 #include "torch/csrc/utils/tuple_parser.h"
-#include "torch/csrc/DynamicTypes.h"
-
-namespace pybind11 { namespace detail {
-
-// handle Python <-> torch::autograd::Function conversions
-template <> struct type_caster<std::shared_ptr<torch::autograd::Function>> {
-public:
-  PYBIND11_TYPE_CASTER(std::shared_ptr<torch::autograd::Function>, _("std::shared_ptr<torch::autograd::Function>"));
-
-  bool load(handle src, bool) {
-    if (!THPFunction_Check(src.ptr())) return false;
-    value = THPFunction_asFunction((THPFunction*)src.ptr());
-    return true;
-  }
-  static handle cast(std::shared_ptr<torch::autograd::Function> src, return_value_policy /* policy */, handle /* parent */) {
-    auto fn = functionToPyObject(src);
-    return handle(fn);
-  }
-};
-
-}} // namespace pybind11::detail
 
 using namespace torch::autograd;
 using torch::TupleParser;
